WebAssemblyText

WebAssemblyText.builtinfuncs

WebAssemblyText.argtypes!

WebAssemblyText.blockinfo

WebAssemblyText.blockparse

WebAssemblyText.codeinfo

WebAssemblyText.declaration

WebAssemblyText.getbuiltins

WebAssemblyText.getimports

WebAssemblyText.itemtype

WebAssemblyText.jl2wat

WebAssemblyText.jl2wat

WebAssemblyText.jlstring2wat

WebAssemblyText.jlstring2wat

WebAssemblyText.jsimportentry

WebAssemblyText.process

WebAssemblyText.restructure

WebAssemblyText.stringify

WebAssemblyText.structure

WebAssemblyText.translate

WebAssemblyText.translate

WebAssemblyText.builtinfuncs — Constant

builtinfuncs: a dict with handwritten .wat of some julia builtins.

Details:

cant use types such as Bool or Nothing or multiple return values wasm does comparison as ints with i32.const 0 meaning false so :(iterate) have to return 0 instead of false when iterator is empty, which is an issue when iterating across zero.

WebAssemblyText.argtypes! — Method

argtypes!(ci::CodeInfo, argtypes::Dict, funcs::Dict, items::Array)

Infer argtypes and update argtypes if items[1] is in funcs.

WebAssemblyText.blockinfo — Method

blockinfo(ssa::Array)

Infer a tree of WebAssembly blocks from ssa and return a BlockInfo struct containing

a goto dict

a list of which blocks each ssa index is a child of.

Details

WebAssembly control instructions (mainly) consist of

blocks: block, loop

branching: br, br_if, return

Notes concerning WebAssembly control instructions:

We do not have Phi nodes (a block can only have one parent)

We do not have gotos/jumps. We can only branch backwards/up the block tree.

However, branching to a block continues at end of that block, which effectively is a forward jump (aka break)

Special case: branching to a loop block continues at start of that block (aka continue)

branching is specified in terms of number of levels up (br 0 goes to current block, br 1 goes to parent block and so on. return is essentialy sugar for br MAX)

So the strategy is to infer a blocktree from arbitrary gotos, insert blocks and translate gotos to branching in terms of levels up the tree.

WebAssemblyText.blockparse — Method

blockparse(str::String)

Initial Meta.parse() on entire input. return funcs and initialize Dicts of argtypes and imports.

Details:

funcs: a dict with function,expression as key,values

argtypes: a dict with with function,argtypes as key,values

imports: a dict with with function,importstring as key,values

WebAssemblyText.codeinfo — Method

codeinfo(func::Symbol, argtypes::Array)

Essentially code_typed() with optimize=false

Details:

Not optimizing complicates type inference, but ultimately simplifies translation becuase the returned CodeInfo is much cleaner.

Also, we have to modifiy CodeInfo by

making sure sure slotnames contain names

making sure slottypes and ssavaluetypes contain types

this has mostly to do with iterator variables

WebAssemblyText.declaration — Method

declaration(cinfo, func, argtypes, Rtype)

Get a wat string with function declaration.

WebAssemblyText.getbuiltins — Method

getbuiltins(ssa::Array)

Get an array of strings with any used builtin .wat functions as specified by the dict builtinfuncs.

WebAssemblyText.getimports — Method

getimports(imports::Dict)

Get a .wat string with any used functions that are not builtins or userdefined.

Details

a few basic are builtin to wasm. these can be translated.

other basic functions are bultin to JavaScripts global Math object, these can be imported

many more are builtin to julia... so they need to be implemented either in .jl, .wat or .js

give warning if the function cant be imported from js Math.

WebAssemblyText.itemtype — Method

itemtype(ci::CodeInfo, item)

Get a concrete type, specialized on item type.

WebAssemblyText.jl2wat — Method

jl2wat(path::AbstractString)

Convert contents of a julia source code file to WebAssembly text.

Examples

julia> using WebAssemblyText
julia> wat = jl2wat("example.jl")
julia> println(wat)

WebAssemblyText.jlstring2wat — Method

jlstring2wat(str::AbstractString)

Convert a string of julia source code to WebAssembly text.

Examples

julia> using WebAssemblyText
julia> str="
hello(x) = 2.0*x
hello(1.0)
";
julia> wat = jlstring2wat(str);
julia> println(wat)

(module 

(func $hello (export "hello") (param $x f32) (result f32)
( return ( f32.mul (f32.const 2.0) (local.get $x) ) ))
)

WebAssemblyText.jsimportentry — Method

jsimportentry(func, argtypes)

Get a string of possible javascript Math module import, assuming it exists in the Math module.

Details

from a func such as sin, return a string like "sin: (x) => Math.sin(x)"

WebAssemblyText.process — Method

process(func, funcs, argtypes)

Get a string with a self contained (func ) expression in .wat format

Details

The main steps of translating a single function

type infer func given argtypes[func]

structure ssa

update argtypes for other functions called in this function

translate to wat and inline to a string

wrap string in a function declaration

WebAssemblyText.restructure — Method

restructure(items)

Restructure items for more straightforward translation.

Details:

expand Nary representation ([mul,a,b,c,d] => [mul,d,[mul,c,[mul,a,b]]])

insert implied iterator increment ([:,1,4] => [:,1,1,4])

pick only initial iterator value if :(iterate) has a single arg [:(iterate),target] => ssa[target][1]

specify :(iteratef) instead of :(iterate) for float iteration.

move ifelse condition to last arg, corresponding to .wat select [ifselse, cond, a, b] => [ifselse, a, b, cond]

TODO:

expand chains of comparisons [comparison, 1, <, i, <=, n] => [&&,[<,1,i],[<=,i,n]]

.wat dont have Bool or Nothing types so iterating across zero does NOT work.. figure out a way to solve this

WebAssemblyText.stringify — Method

inline(ssa::Array)

Get a wat single string from a an already translated ssa.

Details:

Get a list of ssa references

Add parenthesis

Replace refs by the referenced lines

Delete the referenced lines

add block and loop control flow based on GotoNodes

Join.

WebAssemblyText.structure — Method

structure(item::Expr)

Expression as a vector: [head, args...]

Detals:

There are two expression flavors:

head in args[1]

head in head

make them be represented the same way

eval refs of constants

WebAssemblyText.translate — Method

translate(i::Integer, cinfo::CodeInfo, item)

Get a wat string from item, specialized on item type.

WebAssemblyText.translate — Method

translate(ci::CodeInfo, items::Array)

Get a wat string from items, branching to special cases based on items[1].